wip: create CharacterBuilder mvp
I'm commiting this as I may end up using some of it later, but I feel
like I'm railroading myself a direction that may not make sense for this
application. I don't want to lose the work that I've completed, but I
also don't think that this is how the project should start.