feat!: extract Background Builder
Looking over the book again, many of the fields I tied directly to the
Character are derived from the Character's Background. Due to the
growing complexity of the Background struct, I felt it necessary to
create a Builder for it.

I need to be more careful about running main before continuing with a
commit. This commit left main in a broken state, and I didn't catch it
until many commits later...