feat(ancestry): separate entry from instance
This lays the groundwork for tracking Ancestry Traits.

I think need an OriginBuilder as well so I can better enforce Trait
constraints on Characters pulling from multiple Ancestries.

I need to keep the AncestryInstance struct so that the AncestryTraits
stay separated, but I need a better way of tracking which entry the
Traits came from. For now, it's not a priority, but it needs to be a
consideration down the road.