refactor!: offload BackgroundBuilder into macro
I lost a little bit of functionality, but gained back roughly the same
amount. Biggest thing is that I vastly trimmed down that file and gained
the ability to apply the same pattern to the other structs.

Currently, the only "feature" I lost is returning an error when
attempting to build a Background if it doesn't have a Trade or Skill,
but vastly simplifying the API makes it totally worth it.